+++ src/utils/mod.rs
@@ -1,4 +1,5 @@
 pub mod datetime;
 pub mod html;
 pub mod theme;
+pub mod theme_export;
 pub mod url;